refactor: update EnderecoRespDTO ID fields to Long and adjust mapper logic
- Changed `EnderecoRespDTO` ID fields (paisId, ilhaId, concelhoId, freguesiaId, zonaId) from Integer to Long for consistency.
- Updated `EnderecoMapper` to align with the new field type and removed redundant type conversion.
- Adjusted `.igrpstudio` configurations to reflect updated data types.

+++ src/main/java/cv/inps/rh/funcionario/application/dto/EnderecoRespDTO.java
@@ -25,31 +25,31 @@ public class EnderecoRespDTO  {
   private String uuid ;
 
 
-  private Integer paisId ;
+  private Long paisId ;
 
 
   private String paisDesc ;
 
 
-  private Integer ilhaId ;
+  private Long ilhaId ;
 
 
   private String ilhaDesc ;
 
 
-  private Integer concelhoId ;
+  private Long concelhoId ;
 
 
   private String concelhoDesc ;
 
 
-  private Integer freguesiaId ;
+  private Long freguesiaId ;
 
 
   private String freguesiaDesc ;
 
 
-  private Integer zonaId ;
+  private Long zonaId ;
 
 
   private String zonaDesc ;
